% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sankey.R
\name{create_sankey}
\alias{create_sankey}
\title{Sankey plot using ggplot2}
\usage{
create_sankey(data, datacol, pdfname)
}
\arguments{
\item{data}{Input file}

\item{datacol}{Plot parameter}

\item{pdfname}{Output file}
}
\description{
This function creates the sankey plots using the ggplot2 package in R.
More details in the vignettes.
}
\examples{
create_sankey(data,datacol,pdfname)
}
