% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_sankey}
\alias{make_sankey}
\title{Create an object that describes a sankey plot}
\usage{
make_sankey(nodes = NULL, edges)
}
\arguments{
\item{nodes}{A data frame of nodes on the plot, and possibly
their visual style. The first column must be the ids of the
nodes. If this argument is \code{NULL}, then the ids of the
nodes are determined from \code{edges}.}

\item{edges}{A data frame of the edges. The first two columns
must be node ids, and they define the edges. The rest of the columns
contain the visual style of the edges.}
}
\value{
A \code{sankey} object that can be plotted via the
  \code{\link{sankey}} function.x
}
\description{
Create an object that describes a sankey plot
}

