% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R, R/sankey.R
\docType{package}
\name{sankey}
\alias{sankey-package}
\alias{sankey}
\alias{plot.sankey}
\title{Sankey Diagrams}
\usage{
\method{plot}{sankey}(x, ...)

sankey(x, mar = c(0, 5, 0, 5) + 0.2, xlim_expand = 0.2, ylim_expand = 0.1, ...)
}
\arguments{
\item{x}{The plot, created via \code{\link{make_sankey}}.}

\item{...}{Additional arguments, ignored currently.}

\item{mar}{Margin of the plot, see \code{mar} in the \code{par}
manual.}

\item{xlim_expand, ylim_expand}{Range rxpansion factors. A value of 1 means 100% of the range will be added on both sides, effectively tripling the plotting range. (default = \code{0.2}, \code{0.1}).}
}
\value{
Nothing.
}
\description{
Sankey plots illustrate the flow of information or material.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gaborcsardi/sankey#readme}
  \item Report bugs at \url{https://github.com/gaborcsardi/sankey/issues}
}

}
