% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey_diagram.R
\name{sankey_diagram}
\alias{sankey_diagram}
\title{Sankey diagram}
\usage{
sankey_diagram(df, include_borders = TRUE)
}
\arguments{
\item{df}{A dataframe of risk factors over time points}

\item{include_borders}{The option to include white boarders around the alluvium and stratums or no boarders. Either TRUE or FALSE}
}
\description{
Sankey diagram
}
