% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R, R/sankey.R
\docType{package}
\name{sankey}
\alias{plot.sankey}
\alias{sankey}
\alias{sankey-package}
\title{Sankey Diagrams}
\usage{
\method{plot}{sankey}(x, ...)
}
\description{
Sankey plots illustrate the flow of information or material.
}

