% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R, R/sankey.R
\docType{package}
\name{sankey}
\alias{plot.sankey}
\alias{sankey}
\alias{sankey-package}
\title{Sankey Diagrams}
\usage{
\method{plot}{sankey}(x, ...)

sankey(x, mar = c(0, 5, 0, 5) + 0.2, ...)
}
\arguments{
\item{x}{The plot, created via \code{\link{make_sankey}}.}

\item{...}{Additional arguments, ignored currently.}

\item{mar}{Margin of the plot, see \code{mar} in the \code{par}
manual.}
}
\value{
Nothing.
}
\description{
Sankey plots illustrate the flow of information or material.

Draw a sankey plot
}

