\name{sankey_diagram}
\alias{sankey_diagram}
\title{Sankey diagram}
\description{Creates a Sankey diagram to present the vascular risk factor contribution to stroke incidence from years  1990 to 2010.}
\usage{sankey_diagram(df, include_borders = TRUE)}
\arguments{
  \item{df}{A dataframe of 5 risk factors over 5 time points.}
  \item{include_borders}{The option to include white boarders around the alluvium and stratums or no boarders}
}
\value{Returns a Sankey diagram}
\author{Joanne Fil}


